[sub-pixel] Rounding error in table cell height calculation causes unnecessary scrollbar https://bugs.webkit.org/show_bug.cgi?id=111794
Source/WebCore:

Reviewed by Levi Weintraub.

In RenderTableCell::logicalHeightForRowSizing the
adjustedLogicalHeight is calculated from the logicalHeight and
intrinsic padding and is then returned and floored. This can
cause cause the cell to be slightly smaller (0.5px) than the
element it contains.

Test: fast/sub-pixel/table-cell-height.html

* rendering/RenderTableCell.h:
(WebCore::RenderTableCell::logicalHeightForRowSizing):
Change to use pixelSnapped version of logicalHeight method and
change signature to return int as the only caller immediately
down-casts the result to an int.

LayoutTests:

Reviewed by Levi Weintraub.

Add test for zoomed-in table containing block child.

* fast/sub-pixel/table-cell-height-expected.html: Added.
* fast/sub-pixel/table-cell-height.html: Added.


Identifier: 130225@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@145242 268f45cc-cd09-0410-ab3c-d52691b4dbfc