[HTMLTemplateElement] processTemplateEndTag() needs to check for template in html scope https://bugs.webkit.org/show_bug.cgi?id=111880
Reviewed by Eric Seidel.

Source/WebCore:

Currently, </template> handling exits with an error if there is not a template tag "in scope"
which will be true if there is a table (for instance) below a template. This makes it so that
the search (correctly) examines the entire element stack.

Tests added to html5lib testing library.

* html/parser/HTMLTreeBuilder.cpp:
(WebCore::HTMLTreeBuilder::processTemplateEndTag):

LayoutTests:

* html5lib/resources/template.dat:

Identifier: 130263@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@145293 268f45cc-cd09-0410-ab3c-d52691b4dbfc