[chromium] Fix flaky timeouts in platform/chromium/events/fast/events/intercept-postmessage.html https://bugs.webkit.org/show_bug.cgi?id=112089
Patch by James Robinson <jamesr@chromium.org> on 2013-03-11
Reviewed by Adam Barth.

This test had a few problems. The most serious is it raced a window.postMessage() from an inline <script> in the
<head> with the HTML parser processing the rest of the testcase. If the HTML parser yielded before parsing out
the <pre id="console">, function recvMsg1 would generate an error accessing the undefined pre. The test also
set up a setTimeout with a timeout of 50ms to run after a postMessage handler. There's no need to wait after
yielding for a timeout, these events are processed in order.

* platform/chromium/fast/events/intercept-postmessage.html:

Identifier: 130407@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@145473 268f45cc-cd09-0410-ab3c-d52691b4dbfc