[GTK] Enable Web Audio in development builds https://bugs.webkit.org/show_bug.cgi?id=111831
Reviewed by Philippe Normand.

Source/WebKit2:

* GNUmakefile.am: Link the plugin process against libWebCoreModules.la as well.

Tools:

Web Audio is already quite well supported on the GTK port but still has a few issues
that must be addressed before it is enabled by default. Despite that it should be enabled
in development builds just so we are up-to-date with future feature infrastructure changes.
The Web Audio-related layout tests are not run yet due to missing support in DRT/WKTR, but
that's an obvious goal before enabling the feature in release builds.

* Scripts/webkitperl/FeatureList.pm:


Identifier: 130455@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@145526 268f45cc-cd09-0410-ab3c-d52691b4dbfc