Calling WebInspectorClient::highlight() during a fade-out animation of the PageOverlay won't stop its animation https://bugs.webkit.org/show_bug.cgi?id=112271
Reviewed by Tim Horton.

Add a new stopFadeOutAnimation() on PageOverlay that we call when
WebInspectorClient::highlight() is called and a page overlay
is already available. This ensures that any fade-out animation
is cleared before proceeding with showing the page overlay
for the new highlight.

* WebProcess/WebCoreSupport/WebInspectorClient.cpp:
(WebKit::WebInspectorClient::highlight):
* WebProcess/WebPage/PageOverlay.cpp:
(WebKit::PageOverlay::stopFadeOutAnimation):
(WebKit):
* WebProcess/WebPage/PageOverlay.h:
(PageOverlay):


Identifier: 130582@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@145731 268f45cc-cd09-0410-ab3c-d52691b4dbfc