[Texmap] Change brightness filter to match new spec https://bugs.webkit.org/show_bug.cgi?id=112422
New filters spec specifies an intercept of 0 instead of 1.
This is already done for the software version in http://trac.webkit.org/changeset/139770.
This patch adjusts the value for the accelerated version.

Patch by Noam Rosenthal <noam@webkit.org> on 2013-03-15
Reviewed by Allan Sandfeld Jensen.

Filter tests are currently disabled on Qt/GTK/EFL, but they would need to be rebaselined once enabled.

* platform/graphics/texmap/TextureMapperShaderProgram.cpp:

Identifier: 130728@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@145893 268f45cc-cd09-0410-ab3c-d52691b4dbfc