The 'formnovalidate' attribute doesn't work correctly on button elements with child elements https://bugs.webkit.org/show_bug.cgi?id=112541
Patch by Kunihiko Sakamoto <ksakamoto@chromium.org> on 2013-03-18
Reviewed by Kent Tamura.

Source/WebCore:

Fix a bug where formnovalidate attribute of <button> is not honored
if the target of the click event is an element nested within the button.

Test: fast/forms/interactive-validation-formnovalidate-child.html

* html/HTMLFormElement.cpp:
(WebCore::submitElementFromEvent): Looks for the nearest FormControlElement ancestor of the event target.

LayoutTests:

* fast/forms/interactive-validation-formnovalidate-child-expected.txt: Added.
* fast/forms/interactive-validation-formnovalidate-child.html: Added.

Identifier: 130866@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@146055 268f45cc-cd09-0410-ab3c-d52691b4dbfc