[Qt] Monospace font does not render in proper sizes https://bugs.webkit.org/show_bug.cgi?id=93263
Reviewed by Jocelyn Turcotte.

Do not force fonts to use integer metrics. This was necessary when the fast font path
only worked in integer but is now hindering accurate metrics. This is also required to
make kerning in the fast-path look acceptable

Only enabled for Qt 5.1+ to avoid changing all the baselines for 5.0.

* platform/graphics/qt/FontPlatformDataQt.cpp:
(WebCore::FontPlatformData::FontPlatformData):

Identifier: 131004@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@146203 268f45cc-cd09-0410-ab3c-d52691b4dbfc